pathService 리팩토링, MatchMakerService 멀티스레딩 테스트 통과

+++ src/main/java/com/criminals/plusExponential/application/dto/UnmatchedPathDto.java
@@ -21,9 +21,9 @@ public class UnmatchedPathDto {
 
     private int fare;
 
-    private long distance;
+    private int distance;
 
-    private long duration;
+    private int duration;
 
     private User user;
